+++ minos/modules/S7PhysicalHealth.py
@@ -101,7 +101,7 @@ class S7PhysicalHealth(Base):
 
         # Draw individuals next states randomly from this distribution.
         # Update population with new income
-        self.population_view.update(phys_health_prob_df['S7_physical_health'].astype(float))
+        self.population_view.update(phys_health_prob_df['S7_physical_health'].astype(int))
 
     def calculate_S7_physical_health(self, pop):
         """Calculate income transition distribution based on provided people/indices
